packages/shared/requirements.txt
pydantic>=1.10.0,<2.0.0
python-Levenshtein>=0.20.0
aiosqlite>=0.19.0
